:Particle amounts can now be changed in GUI, implemented pause/unpause button

+++ src/Field.cpp
@@ -6,8 +6,8 @@ void Field::spawnRandom(std::vector<Particle> &Particles){
     std::uniform_real_distribution<float> distributionX(p1.x,p2.x); 
     std::uniform_real_distribution<float> distributionY(p1.y, p2.y);
     for(Particle &particle : Particles){
-        int newX = distributionX(gen);
-        int newY = distributionY(gen);
+        float newX = distributionX(gen);
+        float newY = distributionY(gen);
         particle.setPosition(sf::Vector2f(newX, newY));
     }
 }
